moved check to relay.go, and added more info to readme after testing further
